move title and trash logic out of puzzleform
