+++ web/src/components/EditRewardableCell/EditRewardableCell.tsx
@@ -3,7 +3,11 @@ import type {
   EditRewardableMutationVariables,
   FindEditPuzzleQuery,
   FindEditPuzzleQueryVariables,
+  RestoreRewardablePuzzle,
+  RestoreRewardablePuzzleVariables,
   StepType,
+  TrashRewardablePuzzle,
+  TrashRewardablePuzzleVariables,
 } from 'types/graphql'
 
 import { navigate, routes } from '@redwoodjs/router'
@@ -70,6 +74,22 @@ const EDIT_REWARDABLE_MUTATION = gql`
   }
 `
 
+const TRASH_PUZZLE_MUTATION = gql`
+  mutation TrashRewardablePuzzle($rewardableId: String!) {
+    trashRewardablePuzzle(rewardableId: $rewardableId) {
+      success
+    }
+  }
+`
+
+const RESTORE_PUZZLE_MUTATION = gql`
+  mutation RestoreRewardablePuzzle($rewardableId: String!) {
+    restoreRewardablePuzzle(rewardableId: $rewardableId) {
+      success
+    }
+  }
+`
+
 export const Loading = () => <div>Loading...</div>
 
 export const Empty = () => <div>Empty!</div>
@@ -84,7 +104,7 @@ export const Success = ({
   rewardable,
   queryResult,
 }: CellSuccessProps<FindEditPuzzleQuery, FindEditPuzzleQueryVariables>) => {
-  const [editArchetypalPuzzle, { loading, error }] = useMutation<
+  const [editRewardablePuzzle, { loading, error }] = useMutation<
     EditRewardableMutation,
     EditRewardableMutationVariables
   >(EDIT_REWARDABLE_MUTATION, {
@@ -103,6 +123,34 @@ export const Success = ({
       }
 
       return alert('There was an error creating your rewardable!')
+    },
+  })
+
+  const [trashPuzzle] = useMutation<
+    TrashRewardablePuzzle,
+    TrashRewardablePuzzleVariables
+  >(TRASH_PUZZLE_MUTATION, {
+    onCompleted: (data) => {
+      if (
+        data.trashRewardablePuzzle.success &&
+        typeof queryResult?.refetch === 'function'
+      ) {
+        queryResult.refetch()
+      }
+    },
+  })
+
+  const [restorePuzzle] = useMutation<
+    RestoreRewardablePuzzle,
+    RestoreRewardablePuzzleVariables
+  >(RESTORE_PUZZLE_MUTATION, {
+    onCompleted: (data) => {
+      if (
+        data.restoreRewardablePuzzle.success &&
+        typeof queryResult?.refetch === 'function'
+      ) {
+        queryResult.refetch()
+      }
     },
   })
 
@@ -152,44 +200,70 @@ export const Success = ({
       ? dbNft.image
       : ''
 
-  return (
-    <PuzzleForm
-      initialValues={{
-        rewardable: {
-          id: rewardable.id,
-          name: rewardable.name,
-          listPublicly: rewardable.listPublicly,
-          nft: {
-            name: nftName,
-            image: nftImage,
-          },
-        },
-        puzzle: {
-          coverImage: rewardable.puzzle?.coverImage || '',
-          requirements: rewardable.puzzle?.requirements?.length
-            ? rewardable.puzzle.requirements
-            : ['DETAIL'],
-        },
-        steps: formattedSteps,
-      }}
-      refetch={queryResult?.refetch}
-      isEditMode
-      trashed={!!rewardable.trashedAt}
-      onFormSubmit={({ input }) => {
-        if (!rewardable.puzzle?.id) {
-          throw new Error('Error obtaining puzzle ID')
-        }
+  const trashed = !!rewardable.trashedAt
 
-        editArchetypalPuzzle({
-          variables: {
-            input,
-            rewardableId: rewardable.id,
-            puzzleId: rewardable.puzzle.id,
+  return (
+    <div className="form min-h-[calc(100vh-80px)] px-4 pb-16">
+      <div className="my-8 p-2 text-center text-3xl tracking-wide">
+        Edit Your Puzzle
+      </div>
+
+      <button
+        type="button"
+        className="rounded bg-white/10 px-3"
+        onClick={() =>
+          trashed
+            ? restorePuzzle({
+                variables: {
+                  rewardableId: rewardable.id,
+                },
+              })
+            : trashPuzzle({
+                variables: {
+                  rewardableId: rewardable.id,
+                },
+              })
+        }
+      >
+        {trashed ? 'Restore Puzzle' : 'Trash Puzzle'}
+      </button>
+
+      <PuzzleForm
+        initialValues={{
+          rewardable: {
+            id: rewardable.id,
+            name: rewardable.name,
+            listPublicly: rewardable.listPublicly,
+            nft: {
+              name: nftName,
+              image: nftImage,
+            },
           },
-        })
-      }}
-      submissionError={error}
-      submissionPending={loading}
-    />
+          puzzle: {
+            coverImage: rewardable.puzzle?.coverImage || '',
+            requirements: rewardable.puzzle?.requirements?.length
+              ? rewardable.puzzle.requirements
+              : ['DETAIL'],
+          },
+          steps: formattedSteps,
+        }}
+        isEditMode
+        onFormSubmit={({ input }) => {
+          if (!rewardable.puzzle?.id) {
+            throw new Error('Error obtaining puzzle ID')
+          }
+
+          editRewardablePuzzle({
+            variables: {
+              input,
+              rewardableId: rewardable.id,
+              puzzleId: rewardable.puzzle.id,
+            },
+          })
+        }}
+        submissionError={error}
+        submissionPending={loading}
+      />
+    </div>
   )
 }
